Added theme option for pulling list of organizations from CKAN into Wordpress
Added setting "ra_catalogue_api_url" to allow a base URL for a CKAN installation to be added to an RA's settings.  The URL will be used to query the CKAN API and pull down a list of organizations listed in the CKAN server, with paths to logos, site URLs, and proper names (according to selected language).

The list of organizations is written out with links pointing back to the organization page listed on the CKAN install.

Currently, formatted to work with Gutenberg, but could be templated or reformatted to something more generic.